// frameworks/bridge/arkts_frontend/koala_mirror/arkoala-arkts/arkui/src/generated/styled_string.ts
export interface DecorationStyleInterface {
    type: TextDecorationType;
    color?: ResourceColor;
    style?: TextDecorationStyle;
}